feat: enums  SendType
