+++ packages/neuron-ui/src/components/SendFromMultisigDialog/sendFromMultisigDialog.module.scss
@@ -104,6 +104,7 @@
 
 .sendItem {
   & > div {
+    width: 228px;
     input {
       width: 125px !important;
       height: 56px !important;
